+++ src/Release.php
@@ -16,7 +16,7 @@ class Release
     /**
      * Version of the SDK
      */
-    const VERSION = '1.8.1633';
+    const VERSION = '1.8.1634';
 
     /**
      * @param Event $event